Frontend: Add WEC Location Map
The WEC location map displays the current location of the WEC. It uses
react-leaflet which is a wrapper of leaflet. Getting leaflet running on
next.js requires using dynamic components which are called in the index.

As leaflet uses `window` this file cannot live in the `pages` directory.
Using leaflet in this way leads to hydration errors. This may be
solveable but is not necessary at the moment.