Backend: Add handler for spectra files
`SpectraHandler` handles the reading and writing of raw spectra
dataframes. The spectra dataframes have a unique format that is
difficult to store in a database. The workaround here is we store the
output spectra dataframe and calculate the mhkit spectra metrics and
store those in the database.

Of note, spectra files contain one spectra per hour. Additionally, there
is no active archive (this I could find) for spectra files for station
51210.